requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
Flask-CKEditor==0.5.1
python-slugify==8.0.1
beautifulsoup4==4.12.3
gunicorn==21.2.0
python-dotenv==1.0.0
